♻️ [Refactoring]: #184 document/ 低レベル util をカテゴリ別サブフォルダに集約
encoding/ date/ metadata/ catalog/ の 4 つの category 親フォルダを新設し、
旧 document/ 直下の 6 サブフォルダ (pdf-doc-encoding, decode-pdf-string, pdf-date,
document-metadata, document-info-parser, catalog-parser) を git mv で集約。
各 category に barrel (index.ts) を新設し、document/index.ts は category 経由の
再エクスポートに書き換えた。公開 API surface (型 13 + 値 7 = 20 シンボル) は
名前を含めて不変。

Refs #184
