+++ packages/core/src/document/catalog/index.ts
@@ -1,0 +1,2 @@
+export type { ParsedCatalog, ResolveRef } from "./catalog-parser";
+export { CatalogParser } from "./catalog-parser";
