+++ packages/core/src/document/date/index.ts
@@ -1,0 +1,1 @@
+export { parsePdfDate } from "./pdf-date";
