+++ packages/core/src/document/date/pdf-date/__tests__/pdf-date.error.test.ts
