+++ packages/core/src/document/date/pdf-date/index.ts
