+++ packages/core/src/document/encoding/decode-pdf-string/__tests__/decode-pdf-string.decode.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "vitest";
-import type { PdfWarning } from "../../../pdf/errors/warning/index";
-import type { PdfString } from "../../../pdf/types/pdf-types/index";
+import type { PdfWarning } from "../../../../pdf/errors/warning/index";
+import type { PdfString } from "../../../../pdf/types/pdf-types/index";
 import { decodePdfString } from "../../decode-pdf-string";
 import { REPLACEMENT_CHAR } from "../../pdf-doc-encoding";
