+++ packages/core/src/document/encoding/index.ts
@@ -1,0 +1,6 @@
+export { decodePdfString } from "./decode-pdf-string";
+export {
+  decodePdfDocEncoding,
+  PDF_DOC_ENCODING,
+  REPLACEMENT_CHAR,
+} from "./pdf-doc-encoding";
