+++ packages/core/src/document/encoding/pdf-doc-encoding/__tests__/pdf-doc-encoding.decode.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import type { PdfWarning } from "../../../pdf/errors/warning/index";
+import type { PdfWarning } from "../../../../pdf/errors/warning/index";
 import { decodePdfDocEncoding, REPLACEMENT_CHAR } from "../../pdf-doc-encoding";
 
 test("ASCII バイト列をそのままデコードする", () => {
