+++ packages/core/src/document/metadata/document-metadata/__tests__/document-metadata.parse.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "vitest";
-import type { PdfWarning } from "../../../pdf/errors/warning/index";
-import type { PdfValue } from "../../../pdf/types/pdf-types/index";
+import type { PdfWarning } from "../../../../pdf/errors/warning/index";
+import type { PdfValue } from "../../../../pdf/types/pdf-types/index";
 import { PdfTrapped, parseTrappedName } from "../../document-metadata";
 
 const makeName = (value: string): PdfValue => ({ type: "name", value });
