+++ packages/core/src/document/metadata/index.ts
@@ -1,0 +1,4 @@
+export type { ParsedDocumentInfo } from "./document-info-parser";
+export { DocumentInfoParser } from "./document-info-parser";
+export type { DocumentMetadata } from "./document-metadata";
+export { PdfTrapped, parseTrappedName } from "./document-metadata";
